+++ src/lib/FormLayout.php
@@ -6,6 +6,7 @@ interface FormLayout {
 	public function preamble($form);
 	public function postamble($form);
 	public function render_group($group, $res);
+	public function render_hidden($field);
 	public function render_field($field, $error);
 	public function render_fieldset($fieldset);
 	public function render_hint($field);